refactor: move dialer's errorwrapper in i/errorsx
Part of https://github.com/ooni/probe/issues/1505
